Modified perf profiler to allow the raw data file to be copied back to the host. Also added options which allow the call graph collection to be disabled and reporting on the target to be disabled.
BUG=chromium-os:5811
TEST=Used the profiler with a custom autotest and checked that the new
options worked as expected.

Reviewed-on: http://gerrit.chromium.org/gerrit/1770
Tested-by: Raymes Khoury <raymes@chromium.org>

+++ client/profilers/perf/perf.py
@@ -13,7 +13,11 @@ from autotest_lib.client.bin import profiler, os_dep, utils
 class perf(profiler.profiler):
     version = 1
 
-    def initialize(self, events=["cycles","instructions"], trace=False):
+    def initialize(self, events=["cycles","instructions"], trace=False,
+                   delete_data=True, do_report=True, call_graph=True):
+        self.delete_data = delete_data
+        self.do_report = do_report
+        self.call_graph = call_graph
         if type(events) == str:
             self.events = [events]
         else:
@@ -37,7 +41,7 @@ class perf(profiler.profiler):
         self.logfile = os.path.join(test.profdir, "perf")
         cmd = ("exec %s record -a -o %s" %
                (self.perf_bin, self.logfile))
-        if "parent" in self.sort_keys:
+        if "parent" in self.sort_keys and self.call_graph:
             cmd += " -g"
         if self.trace:
             cmd += " -R"
@@ -53,27 +57,29 @@ class perf(profiler.profiler):
 
 
     def report(self, test):
-        for key in self.sort_keys:
-            reportfile = os.path.join(test.profdir, '%s.comm' % key)
-            cmd = ("%s report -i %s --sort %s,dso" % (self.perf_bin,
-                                                      self.logfile,
-                                                      key))
-            outfile = open(reportfile, 'w')
-            p = subprocess.Popen(cmd, shell=True, stdout=outfile,
-                                 stderr=subprocess.STDOUT)
-            p.wait()
+        if self.do_report:
+            for key in self.sort_keys:
+                reportfile = os.path.join(test.profdir, '%s.comm' % key)
+                cmd = ("%s report -i %s --sort %s,dso" % (self.perf_bin,
+                                                          self.logfile,
+                                                          key))
+                outfile = open(reportfile, 'w')
+                p = subprocess.Popen(cmd, shell=True, stdout=outfile,
+                                     stderr=subprocess.STDOUT)
+                p.wait()
 
-        if self.trace:
-            tracefile = os.path.join(test.profdir, 'trace')
-            cmd = ("%s script -i %s" % (self.perf_bin, self.logfile,))
+            if self.trace:
+                tracefile = os.path.join(test.profdir, 'trace')
+                cmd = ("%s script -i %s" % (self.perf_bin, self.logfile,))
 
-            outfile = open(tracefile, 'w')
-            p = subprocess.Popen(cmd, shell=True, stdout=outfile,
-                                 stderr=subprocess.STDOUT)
-            p.wait()
+                outfile = open(tracefile, 'w')
+                p = subprocess.Popen(cmd, shell=True, stdout=outfile,
+                                     stderr=subprocess.STDOUT)
+                p.wait()
 
-        # The raw detailed perf output is HUGE.  We cannot store it by default.
-        perf_log_size = os.stat(self.logfile)[stat.ST_SIZE]
-        logging.info('Removing %s after generating reports (saving %s bytes).',
-                     self.logfile, perf_log_size)
-        os.unlink(self.logfile)
+        # The perf data file is large. Delete it by default, but the user might want it.
+        if self.delete_data:
+            perf_log_size = os.stat(self.logfile)[stat.ST_SIZE]
+            logging.info('Removing %s after generating reports (saving %s bytes).',
+                         self.logfile, perf_log_size)
+            os.unlink(self.logfile)
